pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "memex-twos-mcp"
version = "0.1.0"
description = "MCP server for querying personal task data from Twos app exports"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "PyYAML>=6.0.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "types-PyYAML>=6.0.12.20250915",
]

[project.scripts]
memex-twos-mcp = "memex_twos_mcp.server:main"
